Moved utility classes Moved all the utility classes to the ./Utilities/ folder.
